2 задача расширена до десятичных чисел

+++ 002/Program.cs
@@ -1,8 +1,8 @@
 ﻿//2.Даны два числа. Показать большее и меньшее число
 Console.WriteLine("Введите первое число");
-int n=int.Parse(Console.ReadLine());
+double n=double.Parse(Console.ReadLine());
 Console.WriteLine("Введите второе число");
-int m=int.Parse(Console.ReadLine());
+double m=double.Parse(Console.ReadLine());
 if (n>m)
 {
     Console.WriteLine("Число "+n+" больше числа "+m);
